Refactor Swan service, delete bitcoinreserve
fix tests for services move to specterext and complete
