+++ src/cryptoadvance/specterext/swan/static/swan/css/styles.css
