+++ src/cryptoadvance/specterext/swan/templates/swan/components/base.jinja
