+++ src/cryptoadvance/specterext/swan/templates/swan/components/swan_tab.jinja
