+++ src/cryptoadvance/specterext/swan/templates/swan/index.jinja
