+++ src/cryptoadvance/specterext/swan/templates/swan/oauth2_success.jinja
