+++ src/cryptoadvance/specterext/swan/templates/swan/settings.jinja
